+++ CHANGELOG.md
@@ -1,5 +1,4 @@
-﻿
-<!--
+﻿<!--
 Copyright Glen Knowles 2016 - 2022.
 Distributed under the Boost Software License, Version 1.0.
 -->
@@ -10,6 +9,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org).
+
+## Unreleased (2022-02-14)
+- Changed - cli.exec() returns was work attempted, instead of exit code
 
 ## dimcli 6.2.1 (2022-02-07)
 - Fixed - minWidth of 0 rejected in text columns